project_01/insurance_management/requirements.txt
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
certifi
cffi
charset-normalizer
colorama
comm
debugpy
decorator
defusedxml
Django
django-cors-headers
django-filter
django-watchman
djangorestframework
djangorestframework_simplejwt
exceptiongroup
executing
fastjsonschema
fqdn
h11
httpcore
httpx
ibm_db
ibm_db_django
idna
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
MarkupSafe
matplotlib-inline
mistune
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
overrides
packaging
pandocfilters
parso
platformdirs
prometheus_client
prompt_toolkit
psutil
pure_eval
pycparser
Pygments
PyJWT
python-dateutil
python-decouple
python-dotenv
python-json-logger
pywin32
pywinpty
PyYAML
pyzmq
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rpds-py
Send2Trash
six
sniffio
soupsieve
sqlparse
stack-data
terminado
tinycss2
tomli
tornado
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
